refactor: create bootstrapApp function to make code easier to read
